fix: webpack module error
