+++ .eslintignore
@@ -3,5 +3,6 @@ client/node_modules/**
 client/out/**
 server/node_modules/**
 server/out/**
+*.webpack.config.js
 webpack.config.js
 gulpfile.js
