+++ package.json
@@ -143,8 +143,8 @@
   },
   "scripts": {
     "vscode:prepublish": "npm run package",
-    "compile": "webpack --mode development",
-    "watch": "webpack --mode development --watch",
+    "compile": "webpack --mode development --config ./client/webpack.config.js && webpack --mode development --config ./server/webpack.config.js",
+    "watch": "webpack --mode development --config ./client/webpack.config.js --watch && webpack --mode development --config ./server/webpack.config.js --watch",
     "package": "webpack --mode production --devtool hidden-source-map",
     "lint": "eslint ./client/src ./server/src --ext .ts,.tsx",
     "postinstall": "cd client && npm install && cd ../server && npm install && cd ..",
@@ -161,6 +161,7 @@
     "eslint": "^7.26.0",
     "fancy-log": "^2.0.0",
     "gulp": "^4.0.2",
+    "merge-options": "^3.0.4",
     "mocha": "^8.3.2",
     "ts-loader": "^9.3.0",
     "typescript": "^4.5.5",
